Fix cross domain user error
Google knows when a member is a USER from another G Suite domain. These
are not in the list of users for the target domain, however, so a direct
relationship cannot be established.

This stops the bleeding, but there is a need to add mapped relationships
for members that don't have an entity in the target domain.